requirements.txt
python=3.8
astropy>=4.0
jupyter>=1.0
cython
matplotlib>=3.5
numpy>=1.2
scipy>=1.7
getdist>=1.1.3
pyyaml
joblib
seaborn
openpyxl
GitPython
